+++ patches/src/main/kotlin/app/morphe/patches/youtube/video/voiceovertranslation/YandexVoiceOverTranslationPatch.kt
@@ -31,7 +31,6 @@ private val yandexVoiceOverTranslationResourcePatch = resourcePatch {
             ResourceGroup(
                 "drawable",
                 "dualvot_yt_yandex_vot.xml",
-                "dualvot_yt_yandex_vot_activated.xml",
             )
         )
     }